tests_requirements.txt
docker
pytest
pytest-asyncio
cassandra-driver
testcontainers-redis
testcontainers[cassandra]
testcontainers[mongodb]
